Add resistor R14 between UPDI_MCU and UPDI pin.
If the MCU and programmer get out of sync and each try to drive the line
in different directions, somebody's going to have a bad time. Small
resistor ensures device safety while allowing high-speed bandwidth.
Value is sufficiently low that even if the programmer also has a 470R
on the line the logic levels will still work OK.

See discussion:
https://github.com/SpenceKonde/AVR-Guidance/blob/master/HardwareNotes/UPDISeriesResistors.md